[REF] website: improve field requirement option
Before this commit:

- In the Requirement option of form fields, when a user selected a
comparator (e.g., "Contains" for text fields), entered values in the
condition input, and then re-selected or chose another comparator,
the previously entered values were lost.

- Earlier, when the comparator changed, all dataset values related to
this option (requirementComparator, requirementCondition,
errorMessage, customError) were removed.

After this commit:

- The above issue is fixed. Now, when a user selects a requirement
comparator again, the previously entered values in the condition
input are preserved.

- When the comparator changes, only the requirementComparator dataset
value is updated. Other related dataset values (requirementCondition,
errorMessage, customError) are retained to avoid losing user
input, and they adapt based on the new comparator value.

task-4364528
