+++ addons/website/static/src/builder/plugins/form/form_option_plugin.js
@@ -43,8 +43,6 @@ import { renderToElement } from "@web/core/utils/render";
 import { selectElements } from "@html_editor/utils/dom_traversal";
 import { BuilderAction } from "@html_builder/core/builder_action";
 import { isSmallInteger } from "@html_builder/utils/utils";
-import { localization } from "@web/core/l10n/localization";
-import { formatDate } from "@web/core/l10n/dates";
 import { getParsedDataFor } from "@website/js/utils";
 import { isTargetVisible } from "@html_builder/core/visibility_plugin";
 import { nodeSize } from "@html_editor/utils/position";
@@ -63,11 +61,9 @@ import { nodeSize } from "@html_editor/utils/position";
  * @property { FormOptionPlugin['prepareConditionInputs'] } prepareConditionInputs
  * @property { FormOptionPlugin['setLabelsMark'] } setLabelsMark
  * @property { FormOptionPlugin['clearValidationDataset'] } clearValidationDataset
- * @property { FormOptionPlugin['defaultMessage'] } defaultMessage
  * @property { FormOptionPlugin['fetchModels'] } fetchModels
  */
 
-const { DateTime } = luxon;
 const DEFAULT_EMAIL_TO_VALUE = "info@yourcompany.example.com";
 export class FormOptionPlugin extends Plugin {
     static id = "websiteFormOption";
@@ -85,7 +81,6 @@ export class FormOptionPlugin extends Plugin {
         "prepareConditionInputs",
         "setLabelsMark",
         "clearValidationDataset",
-        "defaultMessage",
         "fetchModels",
     ];
     /** @type {import("plugins").WebsiteResources} */
@@ -169,7 +164,6 @@ export class FormOptionPlugin extends Plugin {
             PropertyAction,
             SetDependencyValueListAction,
             SetCustomErrorMessageAction,
-            SetDefaultErrorMessageAction,
             SetRequirementComparatorAction,
             SetMultipleFilesAction,
             ToggleAllowEmptyAction,
@@ -856,83 +850,6 @@ export class FormOptionPlugin extends Plugin {
         delete fieldEl.dataset.requirementBetween;
         delete fieldEl.dataset.requirementCondition;
     }
-
-    /**
-     * Generates an error message for requirement set on field if validation fails.
-     *
-     * @param {string} [comparator] The method used to form the error message.
-     * @param {string} [condition] The expected value of the field.
-     * @param {string} [between] The maximum date value if the comparator is
-     *      'between' or '!between'.
-     * @returns {string} The default error message.
-     */
-    defaultMessage(comparator, condition, between, type) {
-        if (["substring", "!substring", "domain"].includes(comparator)) {
-            condition = JSON.parse(condition).map(({ requirement_text }) => requirement_text.trim().toLowerCase());
-        }
-        const textMessages = {
-            contains: _t("This field must include keyword %s.", condition),
-            "!contains": _t("This field must not include keyword %s.", condition),
-            substring: _t("This field must contain one of the keyword(s): '%s'", condition),
-            "!substring": _t("This field must not include the keyword(s): '%s'", condition),
-            greater: _t("Invalid: field is not greater than %s.", condition),
-            less: _t("Invalid: field is not less than %s.", condition),
-            "greater or equal": _t("Invalid: field is not greater than or equal to %s.", condition),
-            "less or equal": _t("Invalid: field is not less than or equal to %s.", condition),
-            domain: _t("This field must have one of these email domain(s): %s.", condition),
-        };
-
-        if (condition && textMessages[comparator]) {
-            return textMessages[comparator];
-        }
-
-        if (["date", "datetime"].includes(type)) {
-            const format = type === "date" ? localization.dateFormat : localization.dateTimeFormat;
-            const start = formatDate(DateTime.fromSeconds(parseInt(condition)), { format });
-            const end = formatDate(DateTime.fromSeconds(parseInt(between)), { format });
-
-            const dateMessages = {
-                dateEqual: _t(
-                    "Entered date or time is not correct! It must be %(start)s (%(format)s).",
-                    { start, format }
-                ),
-                "date!equal": _t(
-                    "Entered date or time is not correct! It must not be %(start)s (%(format)s).",
-                    { start, format }
-                ),
-                before: _t(
-                    "Entered date or time is not correct! It must be before %(start)s (%(format)s).",
-                    { start, format }
-                ),
-                after: _t(
-                    "Entered date or time is not correct! It must be after %(start)s (%(format)s).",
-                    { start, format }
-                ),
-                "equal or before": _t(
-                    "Entered date or time is not correct! It must be before or equal to %(start)s (%(format)s).",
-                    { start, format }
-                ),
-                "equal or after": _t(
-                    "Entered date or time is not correct! It must be after or equal to %(start)s (%(format)s).",
-                    { start, format }
-                ),
-                between: _t(
-                    "Entered date or time is not correct! It must be within %(start)s and %(end)s (%(format)s).",
-                    { start, end, format }
-                ),
-                "!between": _t(
-                    "Entered date or time is not correct! It must not be within %(start)s and %(end)s (%(format)s).",
-                    { start, end, format }
-                ),
-            };
-
-            if (condition && dateMessages[comparator]) {
-                return dateMessages[comparator];
-            }
-        }
-
-        return _t("An error has occurred, the form has not been sent.");
-    }
 }
 
 // Form actions
@@ -1481,7 +1398,12 @@ export class SetRequirementComparatorAction extends BuilderAction {
     static id = "setRequirementComparator";
     static dependencies = ["websiteFormOption"];
     apply({ editingElement: fieldEl }) {
-        this.dependencies.websiteFormOption.clearValidationDataset(fieldEl);
+        if (!fieldEl.dataset.requirementComparator) {
+            this.dependencies.websiteFormOption.clearValidationDataset(fieldEl);
+        } else {
+            delete fieldEl.dataset.customError;
+            delete fieldEl.dataset.errorMessage;
+        }
     }
 }
 /**
@@ -1494,36 +1416,13 @@ export class SetRequirementComparatorAction extends BuilderAction {
 export class SetCustomErrorMessageAction extends BuilderAction {
     static id = "setCustomErrorMessage";
     apply({ editingElement: fieldEl }) {
-        if (!fieldEl.dataset.customError) {
-            fieldEl.dataset.customError = true;
-        } else {
-            delete fieldEl.dataset.customError;
-        }
+        fieldEl.dataset.customError = true;
+    }
+    clean({ editingElement: fieldEl }) {
+        delete fieldEl.dataset.customError;
     }
     isApplied({ editingElement: fieldEl }) {
         return fieldEl.dataset.customError;
-    }
-}
-/**
- * Sets the default error message based on the requirement comparator,
- * condition and type of form fields.
- */
-export class SetDefaultErrorMessageAction extends BuilderAction {
-    static id = "setDefaultErrorMessage";
-    static dependencies = ["websiteFormOption"];
-    apply({ editingElement: fieldEl }) {
-        const {
-            requirementComparator: comparator,
-            requirementCondition: condition,
-            requirementBetween: between,
-            type,
-        } = fieldEl.dataset;
-        fieldEl.dataset.errorMessage = this.dependencies.websiteFormOption.defaultMessage(
-            comparator,
-            condition,
-            between,
-            type
-        );
     }
 }
 /**
